+++ include/LTEngine/random/unix_random.hpp
@@ -10,7 +10,7 @@
 namespace LTEngine::Random {
 	class UnixRandom : public Random {
 	public:
-		UnixRandom(u32 bufferSize = 512);
+		UnixRandom(size_t bufferSize = 512);
 		~UnixRandom() override;
 
 		u8 next_u8() override;
@@ -26,7 +26,7 @@ namespace LTEngine::Random {
 		f32 next_f32() override;
 		f64 next_f64() override;
 
-		void nextBytes(u8 *data, u32 length) override;
+		void nextBytes(u8 *data, size_t length) override;
 
 	private:
 		u32 m_bufferSize;
